Allow Arbitrary Metadata for All Content
This change adds a `meta` attribute to the Front Matter section of every type of content supported by `empress-blog`. This attribute is treated as an arbitrary JSON blob by the Ember-Data models, supporting string, array, and/or object formats.

The `meta` attribute is not used by `empress-blog` by default, but it enables developer support for adding metadata to authors, content, pages, and tags. The blueprints have been updated such that all generators will add a `meta` attribute, making it clear to users that it is supported.

Documentation has been updated to mention this feature is available, and some of the default generated content has had different examples of a `meta` attribute added in those markdown files.

I am not sure how to test this, however, since the existing `meta-tests` appears to exclusively assert the output of Front Matter attributes rendered to the DOM, and by default the `meta` attribute is not rendered by `empress-blog` without custom features being added. However, it is clear from running Ember Inspector that the Ember-Data models have correctly parsed each of the `meta` examples added by this change.